+++ src/lib/util.js
@@ -10,7 +10,7 @@ import * as copy from "clipboard-copy";
 const url = require("url");
 
 const getSelectionAsMarkdown = options => {
-  const turndownService = TurndownService(options);
+  let turndownService = TurndownService(options);
 
   if (options.mathjax) {
     turndownService.use(turndownPluginMathJax);
@@ -32,6 +32,57 @@ const getSelectionAsMarkdown = options => {
 
   if (options.reduceListItemPadding) {
     turndownService.use(turndownPluginListItem);
+  }
+
+  if (options.replaceAngleBrackets) {
+    turndownService.escape = function(string) {
+      return (
+        string
+          // Escape backslash escapes!
+          .replace(/\\(\S)/g, "\\\\$1")
+
+          // Escape headings
+          .replace(/^(#{1,6} )/gm, "\\$1")
+
+          // Escape hr
+          .replace(/^([-*_] *){3,}$/gm, function(match, character) {
+            return match.split(character).join("\\" + character);
+          })
+
+          // Escape ol bullet points
+          .replace(/^(\W* {0,3})(\d+)\. /gm, "$1$2\\. ")
+
+          // Escape ul bullet points
+          .replace(/^([^\\\w]*)[*+-] /gm, function(match) {
+            return match.replace(/([*+-])/g, "\\$1");
+          })
+
+          // Escape blockquote indents
+          .replace(/^(\W* {0,3})> /gm, "$1\\> ")
+
+          // Escape em/strong *
+          .replace(/\*+(?![*\s\W]).+?\*+/g, function(match) {
+            return match.replace(/\*/g, "\\*");
+          })
+
+          // Escape em/strong _
+          .replace(/_+(?![_\s\W]).+?_+/g, function(match) {
+            return match.replace(/_/g, "\\_");
+          })
+
+          // Escape code _
+          .replace(/`+(?![`\s\W]).+?`+/g, function(match) {
+            return match.replace(/`/g, "\\`");
+          })
+
+          // Escape link brackets
+          .replace(/[\[\]]/g, "\\$&")
+
+          // Replace angle brackets
+          .replace(/</g, "&lt;")
+          .replace(/>/g, "&gt;")
+      );
+    };
   }
 
   let html = "";